+++ plugins/CryptoPP/src/base64.h
@@ -44,7 +44,7 @@ static const byte asciiToBin64[] =
 
 static const byte binToAscii64[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
 
-char *base64encode(const char *, const int);
+char *base64encode(const char *, const size_t);
 char *base64decode(const char *, size_t *);
 
 string base64encode(const string);
